# docs/api.yml
openapi: "3.1.0"
info:
  title: Grimoire
  version: "1.0.0"
  description: |
    Self-hosted TTRPG library manager.

    All endpoints except `/api/auth/status`, `/api/auth/setup`, and `/api/auth/login`
    require a valid JWT passed as `Authorization: Bearer <token>`, or as a `?token=`
    query parameter (required for browser-embedded images and file downloads).

    **Roles:**
    - `admin` — full access including user management and app settings
    - `gm` — read everything, edit metadata, rescan library, manage campaigns
    - `player` — read-only access
servers:
  - url: http://localhost:9481
    description: Local instance
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tokenParam:
      type: apiKey
      in: query
      name: token
    apiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Campaign:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        owner_id: { type: string, format: uuid }
        owner_display_name: { type: string }
        is_gm_campaign: { type: boolean }
        gm_title: { type: string }
        parent_campaign_id: { type: string, format: uuid, nullable: true }
        system_id: { type: string, format: uuid, nullable: true }
        has_schedule: { type: boolean }
        members: { type: array, items: { $ref: "#/components/schemas/CampaignMember" } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CampaignMember:
      type: object
      properties:
        user_id: { type: string, format: uuid }
        username: { type: string }
        display_name: { type: string, nullable: true }
        status: { type: string, enum: [invited, accepted, declined] }
        character_name: { type: string, nullable: true }
    SessionNote:
      type: object
      properties:
        id: { type: string, format: uuid }
        campaign_id: { type: string, format: uuid }
        session_date: { type: string, format: date, example: "2026-01-15" }
        title: { type: string }
        player_notes: { type: array, items: { type: object } }
        gm_note: { type: object, nullable: true }
...
security:
  - bearerAuth: []
  - tokenParam: []
paths:
...
  /api/campaigns:
    get:
      tags: [campaigns]
      summary: List campaigns
      description: Returns campaigns owned by or with membership for the current user.
      responses:
        "200":
          description: Campaign list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Campaign" }
    post:
      tags: [campaigns]
      summary: Create a campaign
      description: GM or admin role required when `is_gm_campaign` is true.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                is_gm_campaign: { type: boolean, default: false }
                gm_title: { type: string, default: "Game Master" }
                system_id: { type: string, format: uuid, nullable: true }
                parent_campaign_id: { type: string, format: uuid, nullable: true }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
        "403": { description: GM or admin required for GM campaigns }
  /api/campaigns/{id}:
    get:
      tags: [campaigns]
      summary: Get a campaign
      description: Owner, accepted members, or admin.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Campaign with members and resources
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
        "403": { description: Not a member }
        "404": { description: Not found }
    patch:
      tags: [campaigns]
      summary: Update a campaign
      description: Owner or admin.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                gm_title: { type: string }
                system_id: { type: string, nullable: true }
                parent_campaign_id: { type: string, nullable: true }
      responses:
        "200": { description: Updated }
        "403": { description: Not authorised }
        "404": { description: Not found }
    delete:
      tags: [campaigns]
      summary: Delete a campaign
      description: Owner or admin. Deletes all related data.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "403": { description: Not authorised }
        "404": { description: Not found }
  /api/campaigns/{id}/invite:
    post:
      tags: [campaigns]
      summary: Invite a player
      description: Owner or admin. GM campaigns only.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id]
              properties:
                user_id: { type: string, format: uuid }
      responses:
        "201": { description: Invited }
        "400": { description: Not a GM campaign }
        "403": { description: Not authorised }
        "404": { description: User not found }
        "409": { description: Already invited }
  /api/campaigns/{id}/members/{user_id}:
    patch:
      tags: [campaigns]
      summary: Accept/decline invitation or set character name
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: user_id, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [accepted, declined] }
                character_name: { type: string, nullable: true }
      responses:
        "200": { description: Updated }
        "400": { description: Invalid status }
        "403": { description: Not authorised }
        "404": { description: Member not found }
    delete:
      tags: [campaigns]
      summary: Remove a member
      description: Owner, admin, or the member themselves.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: user_id, required: true, schema: { type: string } }
      responses:
        "204": { description: Removed }
        "403": { description: Not authorised }
  /api/campaigns/{id}/eligible-members:
    get:
      tags: [campaigns]
      summary: List users eligible to be invited
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: Eligible users }
        "403": { description: Not authorised }
  /api/campaigns/resources/search:
    get:
      tags: [campaigns]
      summary: Search resources for the resource picker
      parameters:
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: resource_type, schema: { type: string, enum: [book, map, token] } }
        - { in: query, name: limit, schema: { type: integer, default: 30 } }
      responses:
        "200": { description: Resource matches }
  /api/campaigns/{id}/resources:
    get:
      tags: [campaigns]
      summary: List campaign resources
      description: Non-owners only see resources where `shared` is true.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: Resource list }
        "403": { description: Not a member }
    post:
      tags: [campaigns]
      summary: Link a resource to a campaign
      description: Owner or admin.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resource_type, resource_id]
              properties:
                resource_type: { type: string, enum: [book, map, token] }
                resource_id: { type: string, format: uuid }
                shared: { type: boolean, default: false }
      responses:
        "201": { description: Linked }
        "400": { description: Invalid resource type }
        "403": { description: Not authorised }
        "409": { description: Already linked }
  /api/campaigns/{id}/resources/{res_id}:
    patch:
      tags: [campaigns]
      summary: Update resource sharing
      description: Owner or admin.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: res_id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [shared]
              properties:
                shared: { type: boolean }
      responses:
        "200": { description: Updated }
        "403": { description: Not authorised }
        "404": { description: Not found }
    delete:
      tags: [campaigns]
      summary: Unlink a resource
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: res_id, required: true, schema: { type: string } }
      responses:
        "204": { description: Unlinked }
        "403": { description: Not authorised }
  /api/campaigns/{id}/sessions:
    get:
      tags: [campaigns]
      summary: List session notes
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: Session list }
        "403": { description: Not a member }
    post:
      tags: [campaigns]
      summary: Create a session note
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_date]
              properties:
                session_date: { type: string, format: date, example: "2026-01-15" }
                title: { type: string }
      responses:
        "201": { description: Created }
        "400": { description: Invalid date format }
        "403": { description: Not a member }
  /api/campaigns/{id}/sessions/{sid}:
    get:
      tags: [campaigns]
      summary: Get a session with all notes
      description: GM sees `internal_content`; members see only `external_content`.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: sid, required: true, schema: { type: string } }
      responses:
        "200":
          description: Session detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionNote" }
        "403": { description: Not a member }
        "404": { description: Not found }
    patch:
      tags: [campaigns]
      summary: Update session title
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: sid, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
      responses:
        "200": { description: Updated }
        "403": { description: Not a member }
        "404": { description: Not found }
    delete:
      tags: [campaigns]
      summary: Delete a session note
      description: Owner or admin.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: sid, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "403": { description: Not authorised }
  /api/campaigns/{id}/sessions/{sid}/notes/player:
    put:
      tags: [campaigns]
      summary: Save own player note
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: sid, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string }
      responses:
        "200": { description: Saved }
        "403": { description: Not a member }
        "404": { description: Session not found }
  /api/campaigns/{id}/sessions/{sid}/notes/gm:
    put:
      tags: [campaigns]
      summary: Save GM notes
      description: Owner or admin.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: sid, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                internal_content: { type: string }
                external_content: { type: string }
      responses:
        "200": { description: Saved }
        "403": { description: Not authorised }
        "404": { description: Session not found }
  /api/campaigns/{id}/schedule:
    get:
      tags: [campaigns]
      summary: Get campaign schedule
      description: Returns schedule definition and next 10 session dates.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                type: object
                properties:
                  definition: { type: object }
                  next_sessions:
                    type: array
                    items: { type: string, format: date }
    put:
      tags: [campaigns]
      summary: Set or update campaign schedule
      description: Owner or admin. GM campaigns only.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [frequency, days]
              properties:
                frequency: { type: string, enum: [weekly, biweekly, monthly, custom] }
                days:
                  type: array
                  items: { type: integer, minimum: 0, maximum: 6 }
                  description: "Weekday indices: 0=Monday … 6=Sunday"
                time_utc: { type: string, example: "18:00", nullable: true }
                biweekly_reference: { type: string, format: date, nullable: true }
                monthly_week: { type: integer, nullable: true, description: "1–4 or -1 for last" }
                custom_dates:
                  type: array
                  items: { type: string, format: date }
                  nullable: true
      responses:
        "200": { description: Schedule set }
        "403": { description: Not authorised }
    delete:
      tags: [campaigns]
      summary: Remove campaign schedule
      description: Owner or admin.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "204": { description: Removed }
        "403": { description: Not authorised }
  /api/campaigns/{id}/availability:
    get:
      tags: [campaigns]
      summary: Availability chart
      description: Returns availability for the next 10 scheduled sessions.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: Availability data }
        "403": { description: Not a member }
  /api/campaigns/{id}/availability/{date}:
    put:
      tags: [campaigns]
      summary: Set own availability
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: date, required: true, schema: { type: string, format: date } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [available, tentative, unavailable] }
      responses:
        "200": { description: Saved }
        "403": { description: Not a member }
  /api/campaigns/{id}/availability/{date}/cancel:
    put:
      tags: [campaigns]
      summary: Toggle session cancellation
      description: Owner or admin. Marks or unmarks the entire date as cancelled.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: date, required: true, schema: { type: string, format: date } }
      responses:
        "200": { description: Toggled }
        "403": { description: Not authorised }

  # ── Settings ──────────────────────────────────────────────────────────────
...